refactor: moved apps out of embedchain
