+++ embedchain/__init__.py
@@ -2,7 +2,7 @@ import importlib.metadata
 
 __version__ = importlib.metadata.version(__package__ or __name__)
 
-from .embedchain import App  # noqa: F401
-from .embedchain import OpenSourceApp  # noqa: F401
-from .embedchain import PersonApp  # noqa: F401
-from .embedchain import PersonOpenSourceApp  # noqa: F401
+from embedchain.Apps.App import App  # noqa: F401
+from embedchain.Apps.OpenSourceApp import OpenSourceApp  # noqa: F401
+from embedchain.Apps.PersonApp import (PersonApp,  # noqa: F401
+                                       PersonOpenSourceApp)
